Filtering and outlier detection
Starting work on outlier detection and filtering tools
- Added: rolling_window. Most numpy functions can now inherently be
applied against the SMP force data. Pick_surface is currently using this
inside the SMP class. This will cause the SMP class to break if imported
as a standalone. Need to address this in the short term.
- Added: outliers function that uses a rolling window robust z-score to
identify outliers based on a standard deviation threshold and median
absolute deviation. Currently uses subset but we need to discuss if this
should be made generic or not. If not, needs have data type checking
inserted.